RLHF/golang_alignment_5/104857/turn1/modelA/main.go
package modelA

import (
	"fmt"
)

// Define a DataSynchronizer interface
type DataSynchronizer interface {
	SynchronizeData(source, target []map[string]interface{}) ([]map[string]interface{}, error)
}

// Implement a Synchronizer using functional programming principles
type Synchronizer struct{}

func (s Synchronizer) SynchronizeData(source, target []map[string]interface{}) ([]map[string]interface{}, error) {
	// Use immutable functions to process the data
	updatedTarget := applySynchronizationRules(source, target)
	return updatedTarget, nil
}

// Pure function to apply synchronization rules
func applySynchronizationRules(source, target []map[string]interface{}) []map[string]interface{} {
	// Perform data synchronization logic here using map and filter functions
	// For demonstration, let's just remove items from target that exist in source
	sourceKeys := make(map[string]bool)
	for _, item := range source {
		sourceKeys[fmt.Sprint(item)] = true
	}

	return filter(target, func(item map[string]interface{}) bool {
		return !sourceKeys[fmt.Sprint(item)]
	})
}

// Helper function to filter target items
func filter(slice []map[string]interface{}, predicate func(map[string]interface{}) bool) []map[string]interface{} {
	var result []map[string]interface{}
	for _, item := range slice {
		if predicate(item) {
			result = append(result, item)
		}
	}
	return result
}
